config is together
