--- internal/config/root.go
@@ -1,30 +1,0 @@
-package config
-
-import (
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-)
-
-type Root struct {
-	Debug   bool
-	CfgFile string
-}
-
-func (Root) Init(cmd *cobra.Command) error {
-	cmd.PersistentFlags().BoolP("debug", "d", false, "enable debug mode")
-	if err := viper.BindPFlag("debug", cmd.PersistentFlags().Lookup("debug")); err != nil {
-		return err
-	}
-
-	cmd.PersistentFlags().String("config", "", "configuration file path")
-	if err := viper.BindPFlag("config", cmd.PersistentFlags().Lookup("config")); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (s *Root) Set() {
-	s.Debug = viper.GetBool("debug")
-	s.CfgFile = viper.GetString("config")
-}
